Renamed Interfaces, fixed eslinting issue
