+++ src/interfaces/iLoggerService.ts
