+++ src/interfaces/iStartupConfig.ts
